refactor(racing): RandomNumberGenerator의 generate()가 double이 아니라 Integer리턴으로 변경
- Car, Cars, Game, CarsTest, CarTest, UtilTest에서 관련된 내용 수정

+++ src/main/java/racing/domain/Cars.java
@@ -14,7 +14,7 @@ public class Cars {
                 .collect(Collectors.toList());
     }
 
-    public void play(List<Double> randomNumbers) {
+    public void play(List<Integer> randomNumbers) {
         int length = cars.size();
         if (randomNumbers.size() != length) {
             throw new IllegalArgumentException(RANDOM_NUMBER_COUNT_EXCEPTION_MESSAGE);
